Notifications - move toggle to breadcrumbs, fix styling
make the breadcrumbs narrower, use the rest for the notification button,
the badge styling didn't work, replaced with color for the bell when unread

no margins would be nice but we have no such buttons
