+++ app/javascript/components/notification-drawer/helpers.js
@@ -1,11 +1,3 @@
-export const saveNotificationDrawerVisibility = (isDrawerVisible) => {
-  window.localStorage.setItem('miq-notification-drawer-shown', isDrawerVisible ? 'true' : 'false');
-};
-
-export const unreadCountText = function(count) {
-  return sprintf(n__('%d unread notification', '%d unread notifications', count), count);
-};
-
 export const newCountText = function(count) {
   return sprintf(n__('%d New', '%d New', count), count);
 };
